# Added the `esbuild-loader` for JavaScript files to speed up rebuilding the manual and automated tests