--- a/packages/ckeditor5-dev-tests/package.json
+++ b/packages/ckeditor5-dev-tests/package.json
@@ -21,6 +21,7 @@
     "del": "^5.1.0",
     "depcheck": "^1.3.1",
     "dom-combiner": "^0.1.3",
+    "esbuild-loader": "^2.19.0",
     "fs-extra": "^9.0.0",
     "glob": "^7.1.6",
     "istanbul-instrumenter-loader": "^3.0.1",