@pomek 2022-05-30
We decided to abandon adding the `esbuild-loader` into the CKEditor 5 environment.